requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas==1.5.3
numpy==1.23.5
psycopg2-binary
sqlalchemy
azure-keyvault-secrets
azure-identity
requests